requirements.txt
asgiref==3.2.7
Django==3.0.5
django-filter==2.2.0
django-restframework==0.0.1
djangorestframework==3.11.0
Markdown==3.2.1
pytz==2019.3
sqlparse==0.3.1
